refactor timeseries generation
