--- fmriprep_denoise/data/__init__.py
@@ -1,9 +1,0 @@
-from .atlas import fetch_atlas_path, get_atlas_dimensions
-from .preprocess import (get_prepro_strategy, load_phenotype,
-                         load_valid_timeseries, compute_connectome,
-                         check_extraction)
-
-
-__all__ = ['fetch_atlas_path', 'get_atlas_dimensions', 'get_prepro_strategy',
-           'load_phenotype', 'load_valid_timeseries', 'compute_connectome',
-           'check_extraction']
